refactor: replace MakeGroupsPage class with function
