+++ src/components/App/App.js
@@ -2,7 +2,8 @@ import React, { Component } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import MixEdContext from '../../context/MixEdContext';
 import LandingPage from '../../routes/LandingPage/LandingPage';
-import MakeGroupsPage from '../../routes/MakeGroupsPage/MakeGroupsPage';
+// import MakeGroupsPage from '../../routes/MakeGroupsPage/MakeGroupsPage';
+import { MakeGroupsPage } from '../../routes/MakeGroupsPage';
 import GroupsMadePage from '../../routes/GroupsMadePage/GroupsMadePage';
 import SavedGroupsPage from '../../routes/SavedGroupsPage/SavedGroupsPage';
 import NotFoundPage from '../../routes/NotFoundPage/NotFoundPage';
